+++ gradle-metric-schema/src/main/resources/META-INF/gradle-plugins/com.palantir.metric-schema-dashboard.properties
@@ -1,0 +1,1 @@
+implementation-class=com.palantir.metric.schema.gradle.MetricSchemaDashboardPlugin
