--- gradle-metric-schema/src/main/resources/META-INF/gradle-plugins/com.palantir.metric-schema-datadog.properties
@@ -1,1 +1,0 @@
-implementation-class=com.palantir.metric.schema.gradle.MetricSchemaDataDogPlugin
